% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostats.tutorials-package.R
\docType{package}
\name{biostats.tutorials-package}
\alias{biostats.tutorials}
\alias{biostats.tutorials-package}
\title{biostats.tutorials: Apps and tutorials for statistics and data handling in R}
\description{
Biostats tutorials for data handling in R with tidyverse and shiny apps for statstical concepts. Tutorials should appear in the RStudio tutorials tab (usually top right). To get a list of shiny apps, see `?biostats_apps`.
}
\author{
\strong{Maintainer}: Richard Telford \email{richard.telford@uib.no} (\href{https://orcid.org/0000-0001-9826-3076}{ORCID})

}
\keyword{internal}
