% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostats_apps.R, R/f_test_app.R,
%   R/influence_leverage_app.R, R/power_lm_app.R, R/publication_bias_app.R,
%   R/t_test_app.R
\name{biostats_apps}
\alias{biostats_apps}
\alias{f_test_app}
\alias{influence_leverage_app}
\alias{power_lm_app}
\alias{publication_bias_app}
\alias{t_test_app}
\title{Biostats apps}
\usage{
f_test_app()

influence_leverage_app()

power_lm_app()

publication_bias_app()

t_test_app()
}
\description{
Shiny apps developed by BioStats
}
\details{
\subsection{Distributions}{
\itemize{
\item Coming here soon!
}
}

\subsection{Tests}{
\itemize{
\item \code{t_test_app()} shows how the t-distribution is used to test the significance of a test result
\item \code{f_test_app()} shows how the F-distribution is used to test the significance of a test result
}
}

\subsection{Linear models}{
\itemize{
\item \code{influence_leverage_app()} explores the impact of outliers, influential, and leverage points on diagnostic plots for linear models. Click the graph to move the red point and refit the model.
}
}

\subsection{Power tests and publication biases}{
\itemize{
\item \code{publication_bias_app()} explores the impact of publication bias on the apparent strength of the evidence in the published literature.
\item \code{power_lm_app()} runs a power test for a linear model with either a continuous or categorical predictor.
}
}
}
\section{Functions}{
\itemize{
\item \code{f_test_app()}: f-test and f-distribution

\item \code{influence_leverage_app()}: Diagnostic plots of models with outliers, leverage points and influential points

\item \code{power_lm_app()}: Simulate a power test

\item \code{publication_bias_app()}: Simulation of the effect of publication bias

\item \code{t_test_app()}: t-tests and t-distribution

}}
