% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_leverage_app.R
\name{influence_leverage_app}
\alias{influence_leverage_app}
\title{Diagnostic plots with outliers, leverage points and influential points}
\usage{
influence_leverage_app()
}
\description{
Diagnostic plots with outliers, leverage points and influential points
}
