% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_lm_server.R
\name{power_lm_server}
\alias{power_lm_server}
\title{Server for lm power simulation}
\usage{
power_lm_server(input, output)
}
\arguments{
\item{input}{see shinyApp}

\item{output}{see shinyApp}
}
\description{
Server for lm power simulation
}
