% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prerender-tutorial.R
\name{prerender_tutorials}
\alias{prerender_tutorials}
\title{Pre-render tutorials}
\usage{
prerender_tutorials(tutorials, force = FALSE)
}
\arguments{
\item{tutorials}{Vector of tutorials to pre-render. If omitted will pre-render all tutorials in package}

\item{force}{Force pre-rendering when tutorial html files already exists. Default is to skip these for speed.}
}
\description{
Some tutorials fail to run for unknown reasons. If rendered first, they can still work.
}
