% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_bias_server.R
\name{publication_bias_server}
\alias{publication_bias_server}
\title{Server for lm power simulation}
\usage{
publication_bias_server(input, output)
}
\arguments{
\item{input}{see shinyApp}

\item{output}{see shinyApp}
}
\description{
Server for lm power simulation
}
